+++ docs/changelog.md
@@ -3,6 +3,8 @@
 - 🐛 Fix `date` filter issues (#38, #40)
 - ✨ Add `markdownify` for jekyll (#36, #37)
 - ✨ Add `number_of_words` for jekyll
+- ✨ Add jekyll filter `sort`
+- ✨ Add jekyll filter `slugify`
 - ✨ Add `int`, `float`, `str` and `bool` as both filters and globals for all modes (#40)